mouse fix

+++ src/functions.cpp
@@ -15,7 +15,7 @@ int random_number(int min, int max)
 }
 Vector2f plant_place_to_grow(int x, int y)
 {
-    int row = 1, collum = 1;
+    float row = 1, collum = 1;
     while (1)
     {
         if (y <= (row) * 98 + 75)
